chore: organize repo using src
