+++ index.js
@@ -1,9 +1,9 @@
 import { InstanceBase, Regex, runEntrypoint, InstanceStatus } from '@companion-module/base'
-import { getActions } from './actions.js'
-import { getPresets } from './presets.js'
+import { getActions } from './src/actions.js'
+import { getPresets } from './src/presets.js'
 import { getVariables } from './variables.js'
-import { getFeedbacks } from './feedbacks.js'
-import UpgradeScripts from './upgrades.js'
+import { getFeedbacks } from './src/feedbacks.js'
+import UpgradeScripts from './src/upgrades.js'
 import { ChannelType, MonitoredFeedbackType  } from './src/utility/constants.js'
 import { sleep } from './src/utility/helpers.js'
 import { configFields } from './src/config.js'
